+++ lablib_old/__init__.py
@@ -1,13 +1,13 @@
 from . import (
     operators,
     processors,
-    # renderers,
+    renderers,
     utils
 )
 
 __all__ = [
     "operators",
     "processors",
-    # "renderers",
+    "renderers",
     "utils"
 ]
